Change name of method to #list_properties_of
The name of this method was originally #list_elements_with_properties. That would print off a list of all 118 currently known chemical elements and the properties that the program had scraped from Wikipedia. Now, the method instead lists all of the properties of the element that the user wishes to examine.