Fix issues with #list_properties_of
This method was pausing for different lengths of time between certain properties. This was because it was calling #sleep even for values of nil or 0. I fixed that and added spaces between each pair of calls to #puts and #sleep to make the code easier to read. Finally, I deleted a suggestion that I had implemented.