Add walkthrough video and move a sentence elsewhere
The walkthrough video that I added to the Usage section shows how to use the PeriodicTable CLI. I moved the sentence about the NOTES.md file to the Development section as it seems to be more appropriate there. I also made a small edit to the sentence to make the section "flow" better, grammatically speaking.