Combine two methods into one
It turns out that I don't need the #delete_unnecessary_nodes_from method. All I had to do was access scraped_elements[2..-2], which eliminates the unnecessary nodes more quickly and nondestructively. The unnecessary nodes are two thead nodes that Nokogiri put in there by mistake, and the last node in the scraped table, which just contains notes and not chemical elements.