Begin refactoring #list_elements and #display_set_of_ten_elements
To make this program more dynamic, I am going to start eliminating "magic numbers" and allow the user to choose how many elements to view. I will need to make a #page method or something similar for this. I am also going to allow the user to examine an element after seeing a list of elements.