Finish stubbing the additional "examine" option
To finish stubbing the "examine" option for the first ten elements, I first had to fix a few typos. Then, I wrote the Element class method #find_element_by_atomic_number (and realized that I needed to change the atomic_number's value from a string to an integer). This will now examine any of the first ten elements and return an error message if another number is chosen.