Restore original CLI and continue new list feature
I decided to let the CLI work as it did before I started refactoring and adding a new feature, with one exception: it still allows the user to examine one of the first ten elements after viewing a list of elements. I am still working on #list_elements_v2, but it is not yet ready to replace #list_elements. To that end, I figured out how to display any page of elements with any number of elements. Part of that involved adding the #select_elements_by_atomic_number class method to the Element class.