Display blank or zero values as 0 or N/A, respectively
Originally, the CLI wouldn't print any properties that had blank values or whose values equaled zero. Now, it prints all of the elements' properties by printing values of 0 as they are, and printing blank values as "N/A". This required changing the ValueModifier module method #number_or_nil to #number_or_na, and updating a few other methods to reflect this change.