Change method and variable names and add the method #examine_element_from_list
It is now possible for this program to examine an element from a list of elements by specifying its corresponding number in that list. This additional feature required changing the name of #examine_element to #examine_element_from_main_menu to maintain clarity. Code in other methods had to be changed as a result. Additionally, some variable names were changed for clarity.