Fix error in #list_elements
Whenever the user chose to display "all" or "118" of the elements, and if they answered "yes" to examining an element, they would be asked to choose between 0 and 118 instead of 1 and 118. If they chose 0, then the code would attempt to print the properties of the nonexistent Element whose atomic_number is 0, and it would return an error. This is now fixed.